refactor click-to-tweet block
